fixed bug 1 prevent verified

+++ src/controllers/auth-controller.ts
@@ -67,6 +67,11 @@ class UserController {
       return;
     }
 
+    if (user.verified) {
+      res.status(422).json({ error: "Your account is already verified!" });
+      return;
+    }
+
     await EmailVerificationToken.findOneAndDelete({ owner: userId });
 
     const token = generateToken();
